// src/app/models/Textoayuda.model.ts
//import { StringifyOptions } from 'querystring';
import { NodeCompatibleEventEmitter } from 'rxjs/internal/observable/fromEvent';

export class TextoayudaModel {

    title1: string;
    message1: string;
    question1: string;
    response1: string;
    question2: string;
    response2: string;
    question3: string;
    response3: string;
    question4: string;
    response4: string;
    question5: string;
    response5: string;
    noencuentras: string;
    whatsapp: string;
    telefono: string;
    escribenos: string;
    formulario: string;

    constructor(){
            
            this.title1 = "";
            this.message1 = "";
            this.question1 = "";
            this.response1 = "";
            this.question2 = "";
            this.response2 = "";
            this.question3 = "";
            this.response4 = "";
            this.question5 = "";
            this.response5 = "";
            this.noencuentras = "";
            this.whatsapp = "";
            this.telefono = "";
            this.escribenos = "";
            this.formulario = "";


    }

}